requirements.txt
langchain==1.2.15
langchain-core==1.2.29
langchain-text-splitters==1.1.1
langchain-groq==1.1.2
langchain-huggingface==1.0.1
langchain-chroma==1.1.0
chromadb==1.5.7
ragas==0.4.3
python-dotenv==1.2.2
pandas==3.0.2
numpy==2.4.4
sentence-transformers==5.1.2
matplotlib==3.9.2
scipy==1.14.1
